fixed silly bug

+++ src/generator/plugin/c/rand.h
@@ -1,1 +1,1 @@
-real32_t randf(int32_t (*rand)(void*), void*C);
+real32_t randf(int8_t (*rand)(void*), void*C);
